Changed Queue implementation to allow read/read parallelism.
This works by replacing most "synchronized" statements with a
"Reentrant Read-Write Lock". This allows multiple threads to
read the Queue at the same time.

This greatly improves response times and lowers delays when the
server is under high load. Before, all pages that needed to read
the queue for the side-panel were effectively loaded in sequence.