Changed Queue implementation to use re-entrant locks
Certain methods remain synchronized, as the the queueing algorithm
makes certain assumptions about when threads may enter these methods.
Not synchronizing them would not lead to a deadlock, but would cause
threads to "bunch-up" on the queing methods; refusing to do any
sensible work, as well as increasing pressure on the queue locks.